+++ nix/package.nix
@@ -64,7 +64,7 @@ in
 
 stdenvNoCC.mkDerivation {
   pname = "screen-shader";
-  version = "2.0";
+  version = "1.0.1";
 
   dontUnpack = true;
   nativeBuildInputs = [ makeWrapper ];